Another attempt at merging in trunk ... hopefully with proper tracking now.
Something was not right about the earlier merges ... or they did not cover everything. I hope we now got here what was intended.
About the state of the raw decoder as such: The simple test program produces less output as it should. Disabling gapless decoding was not enough. We need to settle that, compliance tests should work.